Rearrange where angle cut off happens
